OR / AND operators to select devices based on tag list
This update allows publisher to choose the logical operator to
process the tag list (of course, it only makes sense when the tag list
has two or more items).

OR: default behaviour. All devices registered to at least one tag on the
list.

AND: Only devices registered simultaneously to all the tags specified on
the list.